requirements.txt
Django==3.0.7
djangorestframework==3.11.0
drf-nested-routers==0.91


